added auth check, root router error page, refactored App.js, exported router to separate file, small style fixes
