+++ src/pages/ErrorPage.js
@@ -1,12 +1,13 @@
 import Box from "@mui/material/Box";
-import Button from "@mui/material/Button";
 import Typography from "@mui/material/Typography";
 import { isRouteErrorResponse, useNavigate, useRouteError } from "react-router-dom";
+import { PrimaryButton } from "../components/PrimaryButton/PrimaryButton";
 
 export const ErrorPage = () => {
   const error = useRouteError();
   const navigate = useNavigate()
   const goBack = () => navigate(-1, { replace: true })
+  const buttonSxProp = { textTransform: 'capitalize', mt: 4, minWidth: 120, flexGrow: 0, boxShadow: 0 }
 
   if (isRouteErrorResponse(error)) {
     return (
@@ -23,9 +24,9 @@ export const ErrorPage = () => {
         <Typography variant='h2' component='h2' gutterBottom>{error?.status}</Typography>
         <Typography variant='body1' component='p'>{error?.statusText}</Typography>
         {error.data?.message && <Typography variant='body2' component='div'>{error.data.message}</Typography>}
-        <Button variant='contained' color="inherit" sx={{ textTransform: 'uppercase', mt: 4, minWidth: 120 }} onClick={goBack}>
+        <PrimaryButton variant='contained' color="inherit" sx={buttonSxProp} onClick={goBack}>
           Go Back
-        </Button>
+        </PrimaryButton>
       </Box>
     );
   } else {
